refactor: Use unsigned long long for think and custom usleep times
Also:
- Remove unnecessary function parameters.
- Refactor `calc_think_us()` to reduce the function length.

+++ philo/include/print_pub.h
@@ -24,7 +24,7 @@ void	print_verbose_monitor(t_philo *philo, const char *msg);
 void	print_actual_elapsed_time(
 			struct timeval *start,
 			struct timeval *end,
-			unsigned int target_time_us,
+			unsigned long long target_time_us,
 			t_philo *me);
 
 #endif
